# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "CIAM Demo Documentation",
...
    "contact": {
      "name": "test",
      "email": "someemail@somedomain.com"
    },
    "version": "v1",
    "x-logo": {
      "url": "https://img.freepik.com/vector-premium/diseno-logotipo-agua-hexagonal-listo-usar_94202-237.jpg?w=360",
      "altText": "Your logo alt text here"
    }
  },
  "paths": {
...
    "/api/Nethone/NethoneTransactionConfirm": {
      "put": {
        "tags": [
          "Nethone"
        ],
        "description": "Validates Nethone transaction integrity.",
        "operationId": "NethoneTransactionConfirm",
        "requestBody": {
          "description": "The Nethone flags to validate.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlagsNethoneConfirm"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NethoneResponseConfirm"
                }
              }
            }
          },
          "404": {
            "description": "Status not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "Payment Required,verify internal log for more information.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized,verify internal log for more information.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Nethone/NethoneTransaction": {
      "post": {
        "tags": [
          "Nethone"
        ],
        "description": "Process Nethone transaction.",
        "operationId": "NethoneTransaction",
        "requestBody": {
          "description": "The Nethone flags to validate.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NethoneTransactionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NethoneResponseConfirm"
                }
              }
            }
          },
          "404": {
            "description": "Status not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "Payment Required,verify internal log for more information.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Nethone/NethoneRecommendation": {
      "post": {
        "tags": [
          "Nethone"
        ],
        "description": "Process Nethone recommendation.",
        "operationId": "NethoneRecommendation",
        "requestBody": {
          "description": "The Nethone flags to validate.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewNethoneRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NethoneResponse"
                }
              }
            }
          },
          "404": {
            "description": "Status not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "Payment Required,verify internal log for more information.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/User/GetLatestTaC": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "This endpoint will get the TaC Version.",
        "operationId": "LatestTaCResponse",
        "parameters": [
          {
            "name": "UserId",
            "in": "query",
            "required": true,
            "style": "form",
            "schema": {
              "type": "string",
              "description": "User Id Auth0",
              "example": "auth0|63b477d3b85540fe7e870500"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LatestTaCResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Batterys": {
        "type": "object",
        "properties": {
          "battery_health": {
            "type": "string",
            "nullable": true
          },
          "extra_health": {
            "type": "integer",
            "format": "int32"
          },
          "levelextra_status": {
            "type": "number",
            "format": "double"
          },
          "extra_level": {
            "type": "integer",
            "format": "int32"
          },
          "extra_scale_level": {
            "type": "integer",
            "format": "int32"
          },
          "battery": {
            "type": "string",
            "nullable": true
          },
          "extra_plugged": {
            "type": "integer",
            "format": "int32"
          },
          "charging_state": {
            "type": "string",
            "nullable": true
          },
          "extra_status": {
            "type": "integer",
            "format": "int32"
          },
          "extra_present": {
            "type": "boolean"
          },
          "extra_technology": {
            "type": "string",
            "nullable": true
          },
          "extra_temperature": {
            "type": "integer",
            "format": "int32"
          },
          "extra_voltage": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "Carrier": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string",
            "nullable": true
          },
          "iso_country_code": {
            "type": "string",
            "nullable": true
          },
          "mobile_country_code": {
            "type": "string",
            "nullable": true
          },
          "mobile_network_code": {
            "type": "string",
            "nullable": true
          },
          "allows_voip": {
            "type": "boolean"
          },
          "icc_card": {
            "type": "boolean"
          },
          "roaming": {
            "type": "boolean"
          },
          "sim_state": {
            "type": "integer",
            "format": "int32"
          },
          "radio_phone_type": {
            "type": "integer",
            "format": "int32"
          },
          "network_type": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DeclaredIp": {
        "type": "object",
        "properties": {
          "addr": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DeviceInside": {
        "type": "object",
        "properties": {
          "android_id": {
            "type": "string",
            "nullable": true
          },
          "model": {
            "type": "string",
            "nullable": true
          },
          "manufacturer": {
            "type": "string",
            "nullable": true
          },
          "brand": {
            "type": "string",
            "nullable": true
          },
          "build_fingerprint": {
            "type": "string",
            "nullable": true
          },
          "product": {
            "type": "string",
            "nullable": true
          },
          "http_proxy": {
            "type": "string",
            "nullable": true
          },
          "wifi_on": {
            "type": "boolean"
          },
          "bt_on": {
            "type": "boolean"
          },
          "uptime_ms": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Devices": {
        "type": "object",
        "properties": {
          "carrier": {
            "$ref": "#/components/schemas/Carrier"
          },
          "disk": {
            "$ref": "#/components/schemas/Disk"
          },
          "device": {
            "$ref": "#/components/schemas/DeviceInside"
          },
          "battery": {
            "$ref": "#/components/schemas/Batterys"
          }
        },
        "additionalProperties": false
      },
      "Disk": {
        "type": "object",
        "properties": {
          "mobile_network_code": {
            "type": "string",
            "nullable": true
          },
          "free_space_root": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "httpCode": {
            "$ref": "#/components/schemas/HttpStatusCode"
          },
          "errorCode": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Fingerprint": {
        "type": "object",
        "properties": {
          "signals": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "profiling_type": {
            "type": "string",
            "nullable": true
          },
          "attempt_reference": {
            "type": "string",
            "nullable": true
          },
          "created_date": {
            "type": "string",
            "nullable": true
          },
          "ip": {
            "type": "string",
            "nullable": true
          },
          "network": {
            "$ref": "#/components/schemas/Network"
          },
          "device": {
            "$ref": "#/components/schemas/Devices"
          }
        },
        "additionalProperties": false
      },
      "FlagsNethoneConfirm": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "flags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FlagsNethoneProperties"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "FlagsNethoneProperties": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "HttpStatusCode": {
        "enum": [
          100,
          101,
          102,
          103,
          200,
          201,
          202,
          203,
          204,
          205,
          206,
          207,
          208,
          226,
          300,
          301,
          302,
          303,
          304,
          305,
          306,
          307,
          308,
          400,
          401,
          402,
          403,
          404,
          405,
          406,
          407,
          408,
          409,
          410,
          411,
          412,
          413,
          414,
          415,
          416,
          417,
          421,
          422,
          423,
          424,
          426,
          428,
          429,
          431,
          451,
          500,
          501,
          502,
          503,
          504,
          505,
          506,
          507,
          508,
          510,
          511
        ],
        "type": "integer",
        "format": "int32"
      },
      "Info": {
        "type": "object",
        "properties": {
          "fingerprint": {
            "$ref": "#/components/schemas/Fingerprint"
          },
          "rules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Rule"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "LatestTaCResponse": {
        "required": [
          "acceptDate",
          "version"
        ],
        "type": "object",
        "properties": {
          "acceptDate": {
            "minLength": 1,
            "type": "string",
            "description": "Date accepted Tac Version",
            "example": "21/03/2023"
          },
          "version": {
            "minLength": 1,
            "type": "string",
            "description": "Version TaC",
            "example": "V1.0"
          }
        },
        "additionalProperties": false
      },
...
      "NethoneResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "id_str": {
            "type": "string",
            "nullable": true
          },
          "advice": {
            "type": "string",
            "nullable": true
          },
          "info": {
            "$ref": "#/components/schemas/Info"
          }
        },
        "additionalProperties": false
      },
      "NethoneResponseConfirm": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "id_str": {
            "type": "string",
            "nullable": true
          },
          "reference": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "NethoneTransactionRequest": {
        "type": "object",
        "properties": {
          "reference": {
            "type": "string",
            "nullable": true
          },
          "merchant_reference": {
            "type": "string",
            "nullable": true
          },
          "payment_method": {
            "type": "string",
            "nullable": true
          },
          "user_reference": {
            "type": "string",
            "nullable": true
          },
          "profiling_reference": {
            "type": "string",
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "nullable": true
          },
          "currency": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "string",
            "nullable": true
          },
          "card_token": {
            "type": "string",
            "nullable": true
          },
          "flags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FlagsNethoneProperties"
            },
            "nullable": true
          },
          "inquiry_id": {
            "type": "integer",
            "format": "int64"
          }
        },
        "additionalProperties": false
      },
      "Network": {
        "type": "object",
        "properties": {
          "link_type": {
            "type": "string",
            "nullable": true
          },
          "user_declared_ip": {
            "$ref": "#/components/schemas/DeclaredIp"
          }
        },
        "additionalProperties": false
      },
      "NewNethoneRequest": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "nullable": true
          },
          "attemptReference": {
            "type": "string",
            "nullable": true
          },
          "inquiryType": {
            "type": "string",
            "nullable": true
          },
          "eventType": {
            "type": "string",
            "nullable": true
          },
          "deviceId": {
            "type": "string",
            "nullable": true
          },
          "requiredEvent": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "Rule": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "id_str": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "advice": {
            "type": "string",
            "nullable": true
          },
          "mode": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}